+++ src/core/simulationModifiers/DeathBox.cpp
@@ -7,7 +7,7 @@
 void FluidSolver::DeathBox::ModifySimulation(FluidSolver::IParticleCollection *collection, SimulationInfo &info) {
 
     #pragma omp parallel for
-    for (uint32_t i = 0; i < collection->GetSize(); i++) {
+    for (int64_t  i = 0; i < collection->GetSize(); i++) {
         auto type = collection->GetParticleType(i);
         if (type == IParticleCollection::ParticleTypeBoundary) {
             continue; // don't calculate unnecessary values for the boundary particles.
